requirements.txt
discord
python-dotenv
unidecode